postgres_to_es/requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
idna==3.4
psycopg2-binary==2.9.5
pydantic==1.10.5
python-dotenv==0.21.1
requests==2.28.2
typing_extensions==4.5.0
urllib3==1.26.14
